+++ lib/minfraud/model/ip_address.rb
@@ -47,21 +47,6 @@ module Minfraud
           end
         end
 
-        # Decorate objects with deprecated attributes and names for backwards
-        # compatibility. Do this here rather than with the overhead of
-        # subclasses/modules for them in the hope that one day we can delete
-        # them.
-
-        # These are named differently in maxmind-geoip2.
-        @country.define_singleton_method(:is_in_european_union) { in_european_union? }
-        @registered_country.define_singleton_method(:is_in_european_union) { in_european_union? }
-        @represented_country.define_singleton_method(:is_in_european_union) { in_european_union? }
-        @traits.define_singleton_method(:is_anonymous) { anonymous? }
-        @traits.define_singleton_method(:is_anonymous_vpn) { anonymous_vpn? }
-        @traits.define_singleton_method(:is_hosting_provider) { hosting_provider? }
-        @traits.define_singleton_method(:is_public_proxy) { public_proxy? }
-        @traits.define_singleton_method(:is_tor_exit_node) { tor_exit_node? }
-
         # Mashify turned each language code into an attribute, but
         # maxmind-geoip2 exposes the names as a hash.
         LANGUAGE_CODES.each do |c|